# feat(core): memoize the normalize function