--- a/packages/core/src/data/__tests__/element-benchmarks.js
+++ b/packages/core/src/data/__tests__/element-benchmarks.js
@@ -1,0 +1,47 @@
+/**
+ * @jest-environment node
+ */
+
+import { Suite } from 'benchmark'
+import I from 'immutable'
+
+import { canonical, isOfKind } from '../'
+
+describe('element benchmarks', () => {
+  const detective = I.fromJS({
+    kind: ['detective', 'the', 'smartest'],
+    name: 'Sherlock Holmes',
+  })
+
+  test.skip('canonical', () => {
+    new Suite()
+      .add('canonical memoized', () => {
+        for (let i = 0; i < 10; i++) {
+          canonical('detective')
+          canonical(['detective', 'the'])
+          canonical(['detective', 'the', 'smartest'])
+          canonical(['detective', 'the', 'smartest', 'ever'])
+        }
+      })
+      .on('cycle', function(event) {
+        console.log(event.target.toString())
+      })
+      .run()
+  })
+
+  test.skip('isOfKind', () => {
+    new Suite()
+      .add('isOfKind memoized', () => {
+        for (let i = 0; i < 10; i++) {
+          isOfKind('detective', detective)
+          isOfKind(['detective', 'the'], detective)
+          isOfKind(['detective', 'the', 'smartest'], detective)
+          isOfKind(['detective', 'the', 'smartest', 'ever'], detective)
+        }
+      })
+      .on('cycle', function(event) {
+        console.log(event.target.toString())
+      })
+      .run()
+  })
+})