--- a/packages/core/src/data/__tests__/element-benchmarks.js
+++ b/packages/core/src/data/__tests__/element-benchmarks.js
@@ -16,12 +16,10 @@ describe('element benchmarks', () => {
   test.skip('canonical', () => {
     new Suite()
       .add('canonical memoized', () => {
-        for (let i = 0; i < 10; i++) {
-          canonical('detective')
-          canonical(['detective', 'the'])
-          canonical(['detective', 'the', 'smartest'])
-          canonical(['detective', 'the', 'smartest', 'ever'])
-        }
+        canonical('detective')
+        canonical(['detective', 'the'])
+        canonical(['detective', 'the', 'smartest'])
+        canonical(['detective', 'the', 'smartest', 'ever'])
       })
       .on('cycle', function(event) {
         console.log(event.target.toString())
@@ -32,12 +30,10 @@ describe('element benchmarks', () => {
   test.skip('isOfKind', () => {
     new Suite()
       .add('isOfKind memoized', () => {
-        for (let i = 0; i < 10; i++) {
-          isOfKind('detective', detective)
-          isOfKind(['detective', 'the'], detective)
-          isOfKind(['detective', 'the', 'smartest'], detective)
-          isOfKind(['detective', 'the', 'smartest', 'ever'], detective)
-        }
+        isOfKind('detective', detective)
+        isOfKind(['detective', 'the'], detective)
+        isOfKind(['detective', 'the', 'smartest'], detective)
+        isOfKind(['detective', 'the', 'smartest', 'ever'], detective)
       })
       .on('cycle', function(event) {
         console.log(event.target.toString())