--- a/packages/core/src/data/__tests__/element-benchmarks.js
+++ b/packages/core/src/data/__tests__/element-benchmarks.js
@@ -5,11 +5,23 @@
 import { Suite } from 'benchmark'
 import I from 'immutable'
 
-import { canonical, isOfKind } from '../'
+import { canonical, isOfKind, kindOf } from '../'
 
 describe('element benchmarks', () => {
-  const detective = I.fromJS({
+  const detective1 = I.fromJS({
+    kind: 'detective',
+    name: 'Sherlock Holmes',
+  })
+  const detective2 = I.fromJS({
+    kind: ['detective', 'the'],
+    name: 'Sherlock Holmes',
+  })
+  const detective3 = I.fromJS({
     kind: ['detective', 'the', 'smartest'],
+    name: 'Sherlock Holmes',
+  })
+  const detective4 = I.fromJS({
+    kind: ['detective', 'the', 'smartest', 'ever'],
     name: 'Sherlock Holmes',
   })
 
@@ -30,10 +42,24 @@ describe('element benchmarks', () => {
   test.skip('isOfKind', () => {
     new Suite()
       .add('isOfKind memoized', () => {
-        isOfKind('detective', detective)
-        isOfKind(['detective', 'the'], detective)
-        isOfKind(['detective', 'the', 'smartest'], detective)
-        isOfKind(['detective', 'the', 'smartest', 'ever'], detective)
+        isOfKind('detective', detective3)
+        isOfKind(['detective', 'the'], detective3)
+        isOfKind(['detective', 'the', 'smartest'], detective3)
+        isOfKind(['detective', 'the', 'smartest', 'ever'], detective3)
+      })
+      .on('cycle', function(event) {
+        console.log(event.target.toString())
+      })
+      .run()
+  })
+
+  test.skip('kindOf', () => {
+    new Suite()
+      .add('kindOf memoized', () => {
+        kindOf(detective1)
+        kindOf(detective2)
+        kindOf(detective3)
+        kindOf(detective4)
       })
       .on('cycle', function(event) {
         console.log(event.target.toString())