--- a/packages/core/src/data/element.js
+++ b/packages/core/src/data/element.js
@@ -2,8 +2,9 @@
 
 import * as R from 'ramda'
 import invariant from 'invariant'
-import deprecated from '../log/deprecated'
 import { List, Seq, is, Iterable } from 'immutable'
+import memoize from '../registry/memoize'
+import deprecated from '../log/deprecated'
 
 /**
  * Checks if a given object is of the provided kind.
@@ -48,7 +49,7 @@ export function isElementRef(obj) {
  * @returns {*}
  */
 export const isExactlyOfKind = R.curry(function isExactlyOfKind(kind, element) {
-  if (element == null || kindOf(element) == null) {
+  if (element == null) {
     return false
   }
 
@@ -122,9 +123,9 @@ export function canonical(ref) {
   return normalize(ref)
 }
 
-function normalize(kind) {
+function _normalize(kind) {
   if (typeof kind === 'string') {
-    return normalize([kind])
+    return List.of(kind)
   }
 
   if (Array.isArray(kind)) {
@@ -141,6 +142,8 @@ function normalize(kind) {
 
   return null
 }
+
+const normalize = memoize(_normalize)
 
 export function pathsToChildElements(element) {
   return childPositions(element).flatMap(childrenPath => {