Resolves #120.

Initial benchmarks:
```
canonical non-memoized x 116,203 ops/sec ±8.39% (69 runs sampled)
isOfKind non-memoized x 61,168 ops/sec ±0.26% (95 runs sampled)

canonical memoized x 512,298 ops/sec ±8.66% (71 runs sampled)
isOfKind memoized x 69,476 ops/sec ±0.56% (86 runs sampled)
```

Check https://github.com/netceteragroup/skele/pull/124#issuecomment-445448152 for updated benchmarks.