@bevkoski 2018-12-08
#### Benchmarks before optimizations
```
canonical non-memoized x 168,937 ops/sec ±0.29% (95 runs sampled)
isOfKind non-memoized x 60,486 ops/sec ±0.27% (93 runs sampled)
kindOf non-memoized x 424,707 ops/sec ±0.37% (94 runs sampled)
```

#### Benchmarks after optimizations
```
canonical memoized x 798,412 ops/sec ±0.40% (92 runs sampled)
isOfKind memoized x 102,094 ops/sec ±0.27% (96 runs sampled)
kindOf memoized x 689,314 ops/sec ±0.40% (94 runs sampled)
```

✔️ improvement of 372% in the `canonical` test
✔️ improvement of 68% in the `isOfKind` test
✔️ improvement of 62% in the `kindOf` test